Preliminary Code Reformat
